src/client/requirements.txt
argparse
paho-mqtt
requests
watchdog